refactor: 수정 api 붙이기 위해 로직 수정
